+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Use the `contributorsNames` index, available in `inventory` since `10.10`. Refs UIIN-1451.
 * Add a warning icon for instance/holdings/item marked as Suppressed from discovery. Refs UIIN-1380.
 * Add a warning icon for instance marked as Staff suppressed. Refs UIIN-1381.
+* Add visual display when item is suppressed from discovery. Refs UIIN-1379.
 * Also support `circulation` `10.0`. Refs UIIN-1488.
 
 ## [6.0.0](https://github.com/folio-org/ui-inventory/tree/v6.0.0) (2021-03-18)